feat: WebView-based plugin editor (React + JUCE 8 WebBrowserComponent)
Replaces the JUCE-native editor with a WebKit-rendered React UI bundled
into BinaryData by esbuild.  C++ ↔ JS communicate via JUCE 8's native
integration (emitEvent / addEventListener).

Bridge correctness fixes that surfaced during integration:
- Parameter IDs use the same `l<n>_<base>` format on both sides
  (was `lane<n>_<base>` on the JS side, so every JS→C++ param write
  silently no-op'd via getParameter() → null).
- `enabled` toggle routes through sendEnabled() — it isn't in LANE_MAP.
- paramChange events from C++ deliver actual (un-normalised) values;
  removed the JS-side double-scaling of choice/int parameters.
- Snapshot crash: scaleRoot/scaleMask are global params, not per-lane,
  so getRawParameterValue("l0_scaleRoot") returned null and the deref
  in sendStateSnapshot() segfaulted.
- Lane palette (color/name) is hydrated from window.LANES on snapshot
  receive — the C++ snapshot rightly doesn't carry visual state, but
  the JS handler was overwriting the demo defaults flat, leaving
  CurvePath with stroke=undefined (path drawn but invisible).
- Range overlay used 8-digit hex (#RRGGBBAA) in SVG fill; WKWebView's
  parser drops the alpha pair and renders opaque.  Now uses fillOpacity.
- Frozen playhead: the C++ 30 Hz timer fires once with phase=0 at
  startup and then never again (engine phase doesn't advance in
  standalone), latching jucePhase to 0 forever.  JUCE phase is now
  honoured only inside a 200 ms freshness window; otherwise the demo
  RAF drives the playhead.
- JS console.log/warn/error mirrored to C++ stderr via a 'log' event
  so we can read WKWebView diagnostics.

+++ Source/SegmentedControl.h
@@ -181,7 +181,7 @@ public:
 
         for (int i = 0; i < n; ++i)
         {
-            const auto sb     = juce::Rectangle<float> (b.getX() + segW * i, b.getY(),
+            const auto sb     = juce::Rectangle<float> (b.getX() + segW * static_cast<float> (i), b.getY(),
                                                         segW, b.getHeight());
             const bool active = (i == _selected);
 
@@ -229,7 +229,7 @@ public:
         if (n == 0) return;
 
         const int  idx       = juce::jlimit (0, n - 1,
-                                             (int) ((float) e.x / (float) getWidth() * n));
+                                             (int) ((float) e.x / (float) getWidth() * static_cast<float> (n)));
         const bool wasAlready = (idx == _selected);
         setSelectedIndex (idx);   // fires onChange only if changed
         if (onTap) onTap (idx, wasAlready);
